requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# LLM SDKs
anthropic>=0.34.0
openai>=1.40.0
google-generativeai>=0.7.0

# PII Detection
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
spacy>=3.7.0

# Rate limiting & caching
redis>=5.0.0
slowapi>=0.1.9

# Dashboard
streamlit>=1.36.0
plotly>=5.22.0
pandas>=2.2.0
requests>=2.32.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
